fix logic error in appendIntOrLL

+++ db/jsobj.h
@@ -1238,12 +1238,12 @@ namespace mongo {
             append( fieldName.c_str() , n );
         }
 
-        /** appends a number.  if n < max(long long)/2 then uses int, otherwise long long */
+        /** appends a number.  if n < max(int)/2 then uses int, otherwise long long */
         void appendIntOrLL( const string& fieldName , long long n ){
             long long x = n;
             if ( x < 0 )
                 x = x * -1;
-            if ( x < ( numeric_limits<long long>::max() / 2 ) )
+            if ( x < ( numeric_limits<int>::max() / 2 ) )
                 append( fieldName.c_str() , (int)n );
             else
                 append( fieldName.c_str() , n );
